src/infrastructure/services/extraction/scripts/requirements.txt
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-docx>=1.0.0
